chore : Preview 위치 변경
